+++ Arex388.Carmine/Extensions/ValidationResultExtensions.cs
@@ -1,7 +1,7 @@
 ﻿namespace FluentValidation.Results;
 
 internal static class ValidationResultExtensions {
-	public static IList<string> ToErrors(
+	public static IReadOnlyList<string> ToErrors(
 		this ValidationResult validationResult) => validationResult.Errors.Select(
 		e => e.ErrorMessage).ToList();
 }