+++ Arex388.Carmine/ResponseBase`TResponse.cs
@@ -11,7 +11,7 @@ public abstract class ResponseBase<TResponse>
 	/// <summary>
 	/// The request's errors, if any.
 	/// </summary>
-	public IList<string> Errors { get; internal set; } = [];
+	public IReadOnlyList<string> Errors { get; internal set; } = [];
 
 	/// <summary>
 	/// The request's status.